ft: support `sanitizeInputBeforeValidation` in tool definitions
