+++ apps/hash-ai-worker-ts/src/activities/create-entities-activity.ts
@@ -6,7 +6,7 @@ import type {
   InferenceState,
 } from "./infer-entities/inference-types";
 import { createEntities } from "./infer-entities/persist-entities/create-entities";
-import type { ProposedEntityCreationsByType } from "./infer-entities/persist-entities/generate-persist-entities-tools";
+import type { ProposedEntityCreationsByType } from "./infer-entities/shared/generate-propose-entities-tools";
 
 export const createEntitiesActivity = async (params: {
   actorId: AccountId;
