+++ apps/hash-ai-worker-ts/src/activities/infer-entities/inference-types.ts
@@ -9,7 +9,7 @@ import type OpenAI from "openai";
 
 import type { DereferencedEntityType } from "../shared/dereference-entity-type";
 import type { PermittedOpenAiModel } from "../shared/openai-client";
-import type { ProposedEntityCreationsByType } from "./persist-entities/generate-persist-entities-tools";
+import type { ProposedEntityCreationsByType } from "./shared/generate-propose-entities-tools";
 
 export type CompletionPayload = Omit<
   OpenAI.ChatCompletionCreateParams,
