+++ apps/hash-ai-worker-ts/src/activities/infer-entities/persist-entities/generate-persist-entities-tools.ts
@@ -1,26 +1,22 @@
 import type { JsonObject } from "@blockprotocol/core";
 import type { VersionedUrl } from "@blockprotocol/type-system";
 import { validateVersionedUrl } from "@blockprotocol/type-system";
-import type {
-  ProposedEntity,
-  ProposedEntitySchemaOrData,
-} from "@local/hash-isomorphic-utils/ai-inference-types";
+import type { ProposedEntitySchemaOrData } from "@local/hash-isomorphic-utils/ai-inference-types";
 import type { Entity } from "@local/hash-subgraph";
 import dedent from "dedent";
 import type { JSONSchema } from "openai/lib/jsonschema";
 
 import type { DereferencedEntityType } from "../../shared/dereference-entity-type";
 import type { LlmToolDefinition } from "../../shared/get-llm-response/types";
+import type {
+  ProposedEntityCreationsByType,
+  ProposeEntitiesToolName,
+} from "../shared/generate-propose-entities-tools";
+import { generateProposeEntitiesTools } from "../shared/generate-propose-entities-tools";
 
-export type FunctionName =
-  | "abandon_entities"
-  | "create_entities"
-  | "update_entities";
-
-export type ProposedEntityCreationsByType = Record<
-  VersionedUrl,
-  ProposedEntity[]
->;
+export type PersistEntitiesToolName =
+  | "update_entities"
+  | ProposeEntitiesToolName;
 
 export type ProposedEntityUpdatesByType = Record<
   VersionedUrl,
@@ -120,81 +116,10 @@ export const generatePersistEntitiesTools = (
     schema: DereferencedEntityType;
     isLink: boolean;
   }[],
-): LlmToolDefinition[] => [
+): LlmToolDefinition<PersistEntitiesToolName>[] => [
+  ...generateProposeEntitiesTools(entityTypes),
   {
-    name: "create_entities" satisfies FunctionName,
-    description: "Create entities inferred from the provided text",
-    inputSchema: {
-      type: "object",
-      properties: entityTypes.reduce<Record<VersionedUrl, JSONSchema>>(
-        (acc, { schema, isLink }) => {
-          acc[schema.$id] = {
-            type: "array",
-            title: `${schema.title} entities to create`,
-            items: {
-              $id: schema.$id,
-              type: "object",
-              title: schema.title,
-              description: schema.description,
-              properties: {
-                entityId: {
-                  description:
-                    "Your numerical identifier for the entity, unique among the inferred entities in this conversation",
-                  type: "number",
-                },
-                ...(isLink
-                  ? {
-                      sourceEntityId: {
-                        description:
-                          "The entityId of the source entity of the link",
-                        type: "number",
-                      },
-                      targetEntityId: {
-                        description:
-                          "The entityId of the target entity of the link",
-                        type: "number",
-                      },
-                    }
-                  : {}),
-                properties: {
-                  description: "The properties to set on the entity",
-                  default: {},
-                  type: "object",
-                  properties: schema.properties,
-                },
-              } satisfies ProposedEntitySchemaOrData,
-              required: [
-                "entityId",
-                "properties",
-                ...(isLink ? ["sourceEntityId", "targetEntityId"] : []),
-              ],
-            },
-          };
-          return acc;
-        },
-        {},
-      ),
-    },
-  },
-  {
-    name: "abandon_entities" satisfies FunctionName,
-    description:
-      "Give up trying to create or update entity, following failures which you cannot correct",
-    inputSchema: {
-      type: "object",
-      properties: {
-        entityIds: {
-          type: "array",
-          title: "The entityIds of the entities to abandon",
-          items: {
-            type: "number",
-          },
-        },
-      },
-    },
-  },
-  {
-    name: "update_entities" satisfies FunctionName,
+    name: "update_entities",
     description: dedent(`
             Update entities inferred from the provided text where you have been advised the entity already exists, 
             using the entityId provided. If you have additional information about properties that already exist,
